Reconcile folder structures; homepage to app
