+++ src/components/DepartureInfo/DepartureInfo.tsx
@@ -1,0 +1,1 @@
+//Sudipta's departures component will go here.
