+++ src/components/TrafficInfo/TrafficInfo.tsx
@@ -1,0 +1,1 @@
+// Roz's traffic info component will go here
