+++ src/components/config.example.ts
@@ -3,3 +3,4 @@
 
 export const WEATHER_API_KEY = "your_openweather_api_key_here";
 export const GOOGLE_MAPS_API_KEY = "your_google_maps_api_key_here";
+export const TRAFIKVERKET_API_KEY = `your_trafikverket_api_key_here`;
