+++ src/types/trafficinfo.ts
@@ -1,0 +1,1 @@
+// traffic info interfaces to go here
