+++ src/types/weather.ts
